Add /_status endpoint and readiness checks
Mark the nginx container as ready to serve traffic only when the backend
webserver responds with a status 200 on a /_status endpoint.

This prevents the case where nginx starts up faster than the webserver
(which is usually always the case) and returns 502 until the backend is
ready.

Label containers in development with :latest.